feat: Add template groups and refactor how templates are processed - closes #49
